pyproject.toml
[project]
name = "localtool"
version = "0.1.0"
requires-python = ">=3.12,<3.14"
dependencies = [
    "cryptography",
    "pillow",
    "pyqt6",
    "pyqt6-webengine>=6.11.0",
]

[project.scripts]
localtool = "localtool.main:main"
email    = "localtool.mailer:run"
base64   = "localtool.tools.base64:run"
color    = "localtool.tools.color:run"
deepseek = "localtool.tools.deepseek:run"
exif     = "localtool.tools.exif:run"
gt       = "localtool.tools.gt:run"
hash     = "localtool.tools.hash:run"
httpd    = "localtool.tools.httpd:run"
ip       = "localtool.tools.ip:run"
ll       = "localtool.tools.ll:run"

[project.entry-points."localtool.tools"]
base64   = "localtool.tools.base64:Base64Tool"
color    = "localtool.tools.color:ColorTool"
deepseek = "localtool.tools.deepseek:DeepSeekTool"
email    = "localtool.mailer:EmailTool"
exif     = "localtool.tools.exif:ExifTool"
gt       = "localtool.tools.gt:GtTool"
hash     = "localtool.tools.hash:WinHashTool"
httpd    = "localtool.tools.httpd:HttpdTool"
ip       = "localtool.tools.ip:IpTool"
ll       = "localtool.tools.ll:LlTool"
